requirements.txt
numpy
scikit-learn
nltk
flask
flask-cors
tqdm
joblib
langdetect
requests
mysql-connector-python
